Added instantaneous PSF and E Field
I’ve added an instantaneous PSF and E-Field. They’re not saved (or
calculated) unless saveInstPsf or saveInstScieField are set to true in
the config file.

I’ve made changes in confParse to allow these new parameters

I’ve edited the SCI module, specifically the calcFocalPlane function to
allow the calculation of the e-field and have switched so it bins
before it calcs the intensity (is this ok Andrew?) otherwise I get
differences in the results (e.g. if I later compare intensity
calculated from E field to intensity calculated in the program).

Updated the simulation module to deal with this.

I’ve only updated on config file for now - though I can also do the
others, or leave, as the relevant parameters will automatically be set
to false.